Updated request model to prevent using deprecated method in Rails 7.2
In order to prevent using a deprecated method in Rails 7.2, an adjustment was made to the serialization of the 'params' in the Notable::Request model. Now the coder is assigned with a positional argument for ActiveRecord versions lower than 7.1, and with a named argument for versions equal or higher than 7.1.